Move duplicate code into common
