+++ packages/atxp-world/package.json
@@ -33,8 +33,8 @@
     "pack:dry": "npm pack --dry-run"
   },
   "dependencies": {
-    "@atxp/client": "0.6.1",
-    "@atxp/common": "0.6.1",
+    "@atxp/client": "0.6.2",
+    "@atxp/common": "0.6.2",
     "bignumber.js": "^9.3.0"
   },
   "peerDependencies": {
